blueti: add bluetooth-compat folder
This introduces the bluetooth-compat folder, which is a folder generated after
the admin-refresh & the driver-select procedure.
The idea is to keep ready the folder which would build the modules required
for the target kernel relevant to bluetooth sub-system only - which required a
config.mk update before the generation of this folder.

Change-Id: Icfcd53f3ecb4c8934e16954073bd6799888c4928
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>